fix(canvas) Remove the old floating menu. - Removed `FloatingInsertMenu` from `DesignPanelRootInner`. - Deleted `FloatingMenu` and some supporting values/functions. - Ported the old browser tests for `FloatingMenu` across to the new   canvas toolbar. - Added a wrapper div to `CanvasToolbarSearch` so that it can be tracked by   the tests.
